requirement.txt
streamlit
Pillow 
opencv-python-headless
torch
tqdm
